Bump to v2.0.0
Removed gulp and replaced with nodemon to reload json file everytime there is a speedtest. (I will just stream it in the future when I get time to do so, this is a quick dirty patch instead)

Added security. All messages that are not from SUDO are ignored! Make sure you replace you userId in the bot.js file!

Removed config.json and instead added token and userId constant in bot.js!

Tested /bash command and it works pretty neat on Windows and Linux alike!

ES6 syntax and beautified code among other things!